+++ src/main/java/com/mallang/blog/application/command/BlogSubscribeCommand.java
@@ -1,0 +1,7 @@
+package com.mallang.blog.application.command;
+
+public record BlogSubscribeCommand(
+        Long subscriberId,
+        Long blogId
+) {
+}
