+++ src/main/java/com/mallang/blog/application/command/BlogUnsubscribeCommand.java
@@ -1,0 +1,7 @@
+package com.mallang.blog.application.command;
+
+public record BlogUnsubscribeCommand(
+        Long subscriberId,
+        Long blogId
+) {
+}
