+++ src/main/java/com/mallang/blog/presentation/request/BlogSubscribeRequest.java
@@ -1,0 +1,11 @@
+package com.mallang.blog.presentation.request;
+
+import com.mallang.blog.application.command.BlogSubscribeCommand;
+
+public record BlogSubscribeRequest(
+        Long blogId
+) {
+    public BlogSubscribeCommand toCommand(Long subscriberId) {
+        return new BlogSubscribeCommand(subscriberId, blogId);
+    }
+}
