+++ src/main/java/com/mallang/blog/presentation/request/BlogUnsubscribeRequest.java
@@ -1,0 +1,11 @@
+package com.mallang.blog.presentation.request;
+
+import com.mallang.blog.application.command.BlogUnsubscribeCommand;
+
+public record BlogUnsubscribeRequest(
+        Long blogId
+) {
+    public BlogUnsubscribeCommand toCommand(Long subscriberId) {
+        return new BlogUnsubscribeCommand(subscriberId, blogId);
+    }
+}
