--- src/main/java/com/mallang/subscribe/application/command/BlogSubscribeCommand.java
@@ -1,7 +1,0 @@
-package com.mallang.subscribe.application.command;
-
-public record BlogSubscribeCommand(
-        Long subscriberId,
-        Long blogId
-) {
-}
