--- src/main/java/com/mallang/subscribe/application/command/BlogUnsubscribeCommand.java
@@ -1,7 +1,0 @@
-package com.mallang.subscribe.application.command;
-
-public record BlogUnsubscribeCommand(
-        Long subscriberId,
-        Long blogId
-) {
-}
