+++ src/test/java/com/mallang/common/ServiceTest.java
@@ -9,9 +9,13 @@ import com.mallang.auth.domain.OauthId;
 import com.mallang.auth.query.dao.MemberDao;
 import com.mallang.blog.application.AboutService;
 import com.mallang.blog.application.BlogService;
+import com.mallang.blog.application.BlogSubscribeService;
 import com.mallang.blog.application.command.OpenBlogCommand;
 import com.mallang.blog.domain.AboutRepository;
 import com.mallang.blog.domain.BlogRepository;
+import com.mallang.blog.domain.subscribe.BlogSubscribeRepository;
+import com.mallang.blog.query.dao.SubscriberDao;
+import com.mallang.blog.query.dao.SubscribingBlogDao;
 import com.mallang.category.application.CategoryService;
 import com.mallang.category.domain.CategoryRepository;
 import com.mallang.category.query.CategoryQueryService;
@@ -39,10 +43,6 @@ import com.mallang.post.domain.visibility.PostVisibilityPolicy.Visibility;
 import com.mallang.post.query.PostQueryService;
 import com.mallang.post.query.PostStarQueryService;
 import com.mallang.post.query.dao.PostSearchDao;
-import com.mallang.subscribe.application.BlogSubscribeService;
-import com.mallang.subscribe.domain.BlogSubscribeRepository;
-import com.mallang.subscribe.query.dao.SubscriberDao;
-import com.mallang.subscribe.query.dao.SubscribingBlogDao;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
